Renamed pkg to build_target
